Prototype for Midi to pixels shader, and potentiometer tests
Potentiometer : using aim constraint with tangent vector up to work even with of-axis rotations of the hand.
Behaves better overall, but it now snaps when contact triggers, bad.
I may need to use the "solve in local space" constraint instead...

Midi 2 pixels shader : geometry pass triangle to line of 1 bit encoded pixels.
Idea is good, cheap and practical to implement.
Sadly VRChat desktop view in VR is not a separate render but a copy of left eye with unknown cropping/resizing.
TODO : test using the stream camera in VR.