Improve CS generation tool layout
Split between Runtime/ components and Editor/ generation script.
Make the generation script remove runtime components in the final build.

Checked that with "Delete Unused GameObject" on, d4rkAvatarOptimizer would merge the controller in one skinned mesh.
For this to work, it is better to have the controller as a child of a container which is toggled and parent constrained around.

Also added a debug toggle on the shader for testing only.