Potentiometer geometry+shader tracking
Added shader and tracking triangle generation for potentiometer.

Using an atan2 between direction vectors encoded in the normal field.
Normal field is rotated by skinning, making it easy to track the rotation.

Position it not actually used ; it could be but normal is simpler.

I tried to find a similar branchless vertex identification scheme as with slider/button.
Nothing seemed practical, so reverted to selecting vectors by matching ids.
It generates a nice array of 6 cmov, which is totally ok.